requirements.txt
fastapi
redis
types-redis
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
faker
faker-commerce